Updated Head AI movement to add a sight modifier so each turret can be adjusted indavidualy
Had to put this in the Head AI script cause the profile seems to change all the turrets when I wanted this slider to be for individuals. I defaulted the Modifier to 1 so turrets don't have to be adjusted if you don't want to. Also its a multiplier so make it < 1 for less range or > 1 for more range.